+++ tests/integration/model_contract/insert_invalid_id.sql
@@ -1,0 +1,1 @@
+insert into {database_schema}.seed values (0, 'Cathy', '2022-03-01');