+++ tests/unit/test_macros.py
@@ -337,6 +337,21 @@ class TestDatabricksMacros(unittest.TestCase):
             value = getattr(template.module, name)(relation)
         return re.sub(r"\s\s+", " ", value)
 
+    def __run_macro2(self, template, name, relation, *args):
+        self.default_context["model"].alias = relation
+
+        def dispatch(macro_name, macro_namespace=None, packages=None):
+            if hasattr(template.module, f"databricks__{macro_name}"):
+                return getattr(template.module, f"databricks__{macro_name}")
+            else:
+                return self.default_context[f"spark__{macro_name}"]
+
+        self.default_context["adapter"].dispatch = dispatch
+
+        value = getattr(template.module, name)(*args)
+        value = re.sub(r"\s\s+", " ", value)
+        return value
+
     def test_macros_load(self):
         self.jinja_env.get_template("adapters.sql")
 
@@ -437,3 +452,331 @@ class TestDatabricksMacros(unittest.TestCase):
         )
 
         del self.var["databricks_skip_optimize"]
+
+    def get_test_relation(self):
+        data = {
+            "path": {
+                "database": "some_database",
+                "schema": "some_schema",
+                "identifier": "some_table",
+            },
+            "type": None,
+        }
+        relation = DatabricksRelation.from_dict(data)
+        return relation
+
+    def test_macros_databricks_constraints_to_dbt(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+        constraint = {"name": "name", "condition": "id > 0"}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint]
+        ).strip()
+        self.assertEquals(r, "[{'name': 'name', 'type': 'check', 'expression': 'id > 0'}]")
+
+        constraint = {"condition": "id > 0"}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint]
+        ).strip()
+        assert "raise_compiler_error" in r
+
+        constraint = {"name": "name", "condition": ""}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint]
+        ).strip()
+        assert "raise_compiler_error" in r
+
+        constraint = {"type": "check", "name": "name", "expression": "id > 0"}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint]
+        ).strip()
+        self.assertEquals(r, "[{'type': 'check', 'name': 'name', 'expression': 'id > 0'}]")
+
+        column = {"name": "col"}
+        constraint = {"name": "name", "condition": "id > 0"}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint], column
+        ).strip()
+        assert "raise_compiler_error" in r
+
+        constraint = {"type": "check", "name": "name", "expression": "id > 0"}
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint], column
+        ).strip()
+        self.assertEquals(r, "[{'type': 'check', 'name': 'name', 'expression': 'id > 0'}]")
+
+        constraint = "not_null"
+        r = self.__run_macro2(
+            template, "databricks_constraints_to_dbt", relation, [constraint], column
+        ).strip()
+        self.assertEquals(r, "[{'type': 'not_null', 'columns': ['col']}]")
+
+    def test_macros_get_model_constraints(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+
+        columns = {
+            "id": {"name": "id", "data_type": "int"},
+            "name": {"name": "name", "data_type": "string"},
+        }
+        model = {
+            "columns": columns,
+            "constraints": [{"type": "not_null", "columns": ["id", "name"]}],
+        }
+
+        r = self.__run_macro2(template, "get_model_constraints", relation, model).strip()
+        expected = "[{'type': 'not_null', 'columns': ['id', 'name']}]"
+        assert expected in r
+
+        self.config["persist_constraints"] = True
+        r = self.__run_macro2(template, "get_model_constraints", relation, model).strip()
+        expected = "[{'type': 'not_null', 'columns': ['id', 'name']}]"
+        assert expected in r
+
+        model["meta"] = {"constraints": [{"type": "foo"}]}
+        r = self.__run_macro2(template, "get_model_constraints", relation, model).strip()
+        expected = "[{'type': 'foo'}]"
+        assert expected in r
+
+        self.config["persist_constraints"] = False
+        r = self.__run_macro2(template, "get_model_constraints", relation, model).strip()
+        expected = "[{'type': 'not_null', 'columns': ['id', 'name']}]"
+        assert expected in r
+
+    def test_macros_get_column_constraints(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+        column = {"name": "id"}
+
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        self.assertEqual(r, "[]")
+
+        column["constraints"] = []
+        self.config["persist_constraints"] = True
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        self.assertEqual(r, "[]")
+
+        column["constraints"] = [{"type": "non_null"}]
+        self.config["persist_constraints"] = True
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        self.assertEqual(r, "[{'type': 'non_null'}]")
+
+        self.config["persist_constraints"] = True
+        column["meta"] = {"constraint": "foo"}
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        assert "raise_compiler_error" in r
+
+        column["meta"] = {"constraint": {"condition": "foo", "name": "name"}}
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        assert "raise_compiler_error" in r
+
+        column["meta"] = {"constraint": "not_null"}
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        self.assertEqual(r, "[{'type': 'not_null', 'columns': ['id']}]")
+
+        self.config["persist_constraints"] = False
+        r = self.__run_macro2(template, "get_column_constraints", relation, column).strip()
+        self.assertEqual(r, "[{'type': 'non_null'}]")
+
+    def test_macros_get_constraint_sql_not_null(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+
+        columns = {
+            "id": {"name": "id", "data_type": "int"},
+            "name": {"name": "name", "data_type": "string"},
+        }
+        model = {"columns": columns}
+
+        def run(constraint, column={}):
+            r = self.__run_macro2(
+                template,
+                "get_constraint_sql",
+                relation,
+                relation,
+                constraint,
+                model,
+                column,
+            ).strip()
+
+            return r
+
+        r = run({"type": "not_null", "columns": ["id", "name"]}).strip()
+        expected = "['alter table `some_database`.`some_schema`.`some_table` change column id set not null ;', 'alter table `some_database`.`some_schema`.`some_table` change column name set not null ;']"  # noqa: E501
+        assert expected in r
+
+        r = run({"type": "not_null"}, columns["id"]).strip()
+        expected = "['alter table `some_database`.`some_schema`.`some_table` change column id set not null ;']"  # noqa: E501
+        assert expected in r
+
+        r = run({"type": "not_null", "columns": ["name"]}, columns["id"]).strip()
+        expected = "['alter table `some_database`.`some_schema`.`some_table` change column name set not null ;']"  # noqa: E501
+        assert expected in r
+
+    def test_macros_get_constraint_sql_check(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+
+        columns = {
+            "id": {"name": "id", "data_type": "int"},
+            "name": {"name": "name", "data_type": "string"},
+        }
+        model = {"columns": columns}
+
+        def run(constraint, column={}):
+            r = self.__run_macro2(
+                template,
+                "get_constraint_sql",
+                relation,
+                relation,
+                constraint,
+                model,
+            ).strip()
+
+            return r
+
+        constraint = {
+            "type": "check",
+            "expression": "id != name",
+            "name": "myconstraint",
+            "columns": ["id", "name"],
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint check (id != name);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "check",
+            "expression": "id != name",
+            "name": "myconstraint",
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint check (id != name);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "check",
+            "expression": "id != name",
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint None check (id != name);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "check",
+            "expression": "",
+            "name": "myconstraint",
+        }
+        r = run(constraint)
+        assert "raise_compiler_error" in r
+
+    def test_macros_get_constraint_sql_primary_key(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+
+        columns = {
+            "id": {"name": "id", "data_type": "int"},
+            "name": {"name": "name", "data_type": "string"},
+        }
+        model = {"columns": columns}
+
+        def run(constraint, column={}):
+            r = self.__run_macro2(
+                template,
+                "get_constraint_sql",
+                relation,
+                relation,
+                constraint,
+                model,
+                column,
+            ).strip()
+
+            return r
+
+        constraint = {
+            "type": "primary_key",
+            "name": "myconstraint",
+            "columns": ["name"],
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint primary key(name);']"  # noqa: E501
+        assert expected in r
+
+        column = {"name": "id"}
+        r = run(constraint, column)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint primary key(name);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "primary_key",
+            "name": "myconstraint",
+        }
+        r = run(constraint, column)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint primary key(id);']"  # noqa: E501
+        assert expected in r
+
+    def test_macros_get_constraint_sql_foreign_key(self):
+        template = self.__get_template("adapters.sql")
+        relation = self.get_test_relation()
+
+        columns = {
+            "id": {"name": "id", "data_type": "int"},
+            "name": {"name": "name", "data_type": "string"},
+        }
+        model = {"columns": columns}
+
+        def run(constraint, column={}):
+            r = self.__run_macro2(
+                template,
+                "get_constraint_sql",
+                relation,
+                relation,
+                constraint,
+                model,
+                column,
+            ).strip()
+
+            return r
+
+        constraint = {
+            "type": "foreign_key",
+            "name": "myconstraint",
+            "columns": ["name"],
+            "parent": "parent_table",
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint foreign key(name) references some_schema.parent_table;']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "foreign_key",
+            "name": "myconstraint",
+            "columns": ["name"],
+            "parent": "parent_table",
+            "parent_columns": ["parent_name"],
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint foreign key(name) references some_schema.parent_table(parent_name);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "foreign_key",
+            "name": "myconstraint",
+            "columns": ["name", "id"],
+            "parent": "parent_table",
+            "parent_columns": ["parent_name", "parent_id"],
+        }
+        r = run(constraint)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint foreign key(name, id) references some_schema.parent_table(parent_name, parent_id);']"  # noqa: E501
+        assert expected in r
+
+        constraint = {
+            "type": "foreign_key",
+            "name": "myconstraint",
+            "parent": "parent_table",
+            "parent_columns": ["parent_name"],
+        }
+        column = {"name": "id"}
+        r = run(constraint, column)
+        expected = "['alter table `some_database`.`some_schema`.`some_table` add constraint myconstraint foreign key(id) references some_schema.parent_table(parent_name);']"  # noqa: E501
+        assert expected in r